DAOS-2424 server: Introduce raft_time_t
Currently, the library implements the election timer using the time
duration parameter msec_elapsed of raft_periodic. While this avoids
introducing clocks into the library and eases testing, reasoning
carefully about time becomes harder. Consider the following events on a
follower:

  Time  Event
    10  raft_periodic(...)
    15  raft_recv_appendentries(...)
          // This call sets timeout_elapsed to 0.
    20  raft_periodic(..., 20 - 10 /* msec_elapsed */)
          // The actual time elapsed since the
          // raft_recv_appendentries call is 20 - 15,
          // rather than 20 - 10.

The election timer is _effectively_ reset by the raft_periodic call at
time 10, instead of the raft_recv_appendentries call at time 15. When we
implement leadership leases based on election timers, the lease granted
by this follower effectively starts at time 10!

One can argue that if the user calls raft_periodic with a period way
smaller than the election timeout, the difference between time 10 and
time 15 in the example may not matter to the election timeout or the
lease. For certain users it may be hard to guarantee that the period
will always be small.

Because leadership leases affect not only aliveness but also safety
(i.e., a read sees the latest applied writes), it seems better to
simplify the reasoning by introducing a user-implemented clock and base
election timers on timestamps. The example will become:

  Time  Event
    10  raft_periodic(...)
          // The time of this call no longer affects the
          // following calls.
    15  raft_recv_appendentries(...)
          // This call gets the current time 15 and saves it.
    20  raft_periodic(...)
          // This call gets the current time 20 and computes
          // the time elapsed with 20 - 15.

Signed-off-by: Li Wei <wei.g.li@intel.com>