DAOS-14443 server: Honor lease when campaigning
As a safety measure, raft_election_start should return an error if there
might be a valid lease to another replica. This patch extracts such as
lease check into a common function and changes raft_election_start to
return RAFT_ERR_MIGHT_VIOLATE_LEASE when the lease check is positive.

Signed-off-by: Li Wei <wei.g.li@intel.com>